refactor: Split `src/db.rs` into `src/db/` submodules
Refactors the `GallifreyDB` "God Object" into a modular structure to improve maintainability and cohesion.

- Moves `GallifreyDB` struct definition to `src/db/mod.rs`.
- Splits implementation methods into:
  - `src/db/config.rs`: Configuration and setup.
  - `src/db/transaction.rs`: Transaction management.
  - `src/db/ops.rs`: Basic graph operations.
  - `src/db/temporal.rs`: Temporal query operations.
  - `src/db/vector.rs`: Vector index operations.
  - `src/db/query.rs`: Query planner integration.
  - `src/db/admin.rs`: Admin and maintenance tasks.
- Moves tests to `src/db/tests.rs`.
- Updates `src/lib.rs` to re-export `GallifreyDB` from the new module.

This change does not alter any public API behavior.
