+++ .jules/atlas.md
@@ -32,3 +32,17 @@
 ## 2026-02-01 - Consolidating Versioning
 **Tangle:** Versioning logic was split between `core::version` (metadata) and `storage::version` (data), creating a false boundary and confusion. `storage` exported core domain primitives like `NodeVersion`.
 **Blueprint:** Consolidated all versioning logic into `src/core/version.rs`. Updated `storage` to re-export `core::version` for backward compatibility. This strengthens `core` as the domain definition and `storage` as the implementation.
+
+## 2024-05-23 - Splitting the God Object in src/db.rs
+**Tangle:** `src/db.rs` was a 3500-line "God Object" responsible for everything from configuration and transaction management to vector indexing, temporal queries, and admin operations. This violated the Single Responsibility Principle and made navigation difficult.
+**Blueprint:** Refactored `src/db.rs` into a `src/db/` module directory.
+1. Kept the core `GallifreyDB` struct definition in `mod.rs`.
+2. Extracted implementations into cohesive submodules:
+   - `config.rs`: Initialization and configuration.
+   - `transaction.rs`: Transaction lifecycle management.
+   - `ops.rs`: Basic CRUD and graph operations.
+   - `temporal.rs`: Temporal query operations.
+   - `vector.rs`: Vector index management and search.
+   - `query.rs`: Query builder and executor integration.
+   - `admin.rs`: Maintenance, statistics, and persistence.
+   - `tests.rs`: Unit tests.
