+++ src/db/ops.rs
@@ -1,0 +1,133 @@
+use crate::api::transaction::WriteOps;
+use crate::core::graph::{Edge, Node};
+use crate::core::id::{EdgeId, NodeId};
+use crate::core::property::PropertyMap;
+use crate::db::GallifreyDB;
+use crate::utils::error::Result;
+
+impl GallifreyDB {
+    /// Create a node with the given label and properties.
+    ///
+    /// This is a convenience method that internally uses a write transaction.
+    /// For multiple operations, prefer using `write()` or `write_transaction()`.
+    ///
+    /// # Example
+    ///
+    /// ```ignore
+    /// use gallifreydb::PropertyMapBuilder;
+    ///
+    /// let node_id = db.create_node(
+    ///     "Person",
+    ///     PropertyMapBuilder::new()
+    ///         .insert("name", "Alice")
+    ///         .build()
+    /// )?;
+    /// ```
+    pub fn create_node(&self, label: &str, properties: PropertyMap) -> Result<NodeId> {
+        self.write(|tx| tx.create_node(label, properties))
+    }
+
+    /// Create an edge between two nodes.
+    ///
+    /// This is a convenience method that internally uses a write transaction.
+    /// For multiple operations, prefer using `write()` or `write_transaction()`.
+    ///
+    /// # Example
+    ///
+    /// ```ignore
+    /// use gallifreydb::PropertyMapBuilder;
+    ///
+    /// let edge_id = db.create_edge(
+    ///     source_id,
+    ///     target_id,
+    ///     "KNOWS",
+    ///     PropertyMapBuilder::new().insert("since", 2024).build()
+    /// )?;
+    /// ```
+    pub fn create_edge(
+        &self,
+        source: NodeId,
+        target: NodeId,
+        label: &str,
+        properties: PropertyMap,
+    ) -> Result<EdgeId> {
+        self.write(|tx| tx.create_edge(source, target, label, properties))
+    }
+
+    /// Get the current state of a node.
+    ///
+    /// This uses the fast path (current storage) for O(1) lookup.
+    pub fn get_node(&self, node_id: NodeId) -> Result<Node> {
+        self.current.get_node(node_id)
+    }
+
+    /// Get the current state of an edge.
+    pub fn get_edge(&self, edge_id: EdgeId) -> Result<Edge> {
+        self.current.get_edge(edge_id)
+    }
+
+    // ========================================================================
+    // Zero-copy access methods (Issue #190)
+    // ========================================================================
+
+    /// Get the target node of an edge without cloning the entire edge.
+    ///
+    /// # Performance
+    ///
+    /// - **Zero-copy**: Only reads and returns the target NodeId (8 bytes)
+    /// - **No allocation**: Does not clone Edge or PropertyMap
+    #[inline]
+    pub fn get_edge_target(&self, edge_id: EdgeId) -> Result<NodeId> {
+        self.current.get_edge_target(edge_id)
+    }
+
+    /// Get the source node of an edge without cloning the entire edge.
+    ///
+    /// # Performance
+    ///
+    /// - **Zero-copy**: Only reads and returns the source NodeId (8 bytes)
+    /// - **No allocation**: Does not clone Edge or PropertyMap
+    #[inline]
+    pub fn get_edge_source(&self, edge_id: EdgeId) -> Result<NodeId> {
+        self.current.get_edge_source(edge_id)
+    }
+
+    /// Get outgoing edges from a node (current state).
+    pub fn get_outgoing_edges(&self, node_id: NodeId) -> Vec<EdgeId> {
+        self.current.get_outgoing_edges(node_id)
+    }
+
+    /// Get incoming edges to a node (current state).
+    pub fn get_incoming_edges(&self, node_id: NodeId) -> Vec<EdgeId> {
+        self.current.get_incoming_edges(node_id)
+    }
+
+    /// Get outgoing edges with a specific label (current state).
+    pub fn get_outgoing_edges_with_label(&self, node_id: NodeId, label: &str) -> Vec<EdgeId> {
+        self.current.get_outgoing_edges_with_label(node_id, label)
+    }
+
+    /// Get the number of nodes in the current state.
+    #[inline]
+    pub fn node_count(&self) -> usize {
+        self.current.node_count()
+    }
+
+    /// Get the number of edges in the current state.
+    #[inline]
+    pub fn edge_count(&self) -> usize {
+        self.current.edge_count()
+    }
+
+    /// Get the out-degree of a node (current state).
+    #[inline]
+    pub fn out_degree(&self, node_id: NodeId) -> usize {
+        self.current.out_degree(node_id)
+    }
+
+    /// Get the in-degree of a node (current state).
+    #[inline]
+    pub fn in_degree(&self, node_id: NodeId) -> usize {
+        self.current.in_degree(node_id)
+    }
+}
