+++ src/db/query.rs
@@ -1,0 +1,248 @@
+use crate::core::id::NodeId;
+use crate::core::temporal::Timestamp;
+use crate::db::GallifreyDB;
+use crate::query::builder::state::Initial;
+use crate::query::{Query, QueryBuilder, QueryExecutor, QueryPlanner, QueryResults};
+use crate::utils::error::Result;
+use std::sync::Arc;
+
+impl GallifreyDB {
+    /// Create a new query builder for constructing hybrid queries.
+    ///
+    /// This is the entry point for the fluent query API that enables
+    /// combining graph traversal, vector search, and temporal queries.
+    ///
+    /// # Example
+    ///
+    /// ```ignore
+    /// // Graph + Vector: "Who does Alice know that's similar to Bob?"
+    /// let results = db.query()
+    ///     .start(alice_id)
+    ///     .traverse("KNOWS")
+    ///     .rank_by_similarity(&bob_embedding, 10)
+    ///     .build();
+    ///
+    /// let results = db.execute_query(query)?;
+    ///
+    /// // Temporal + Vector: "What was similar in 2023?"
+    /// let query = db.query()
+    ///     .as_of(timestamp_2023, tx_time)
+    ///     .find_similar(&embedding, 10)
+    ///     .build();
+    /// ```
+    #[must_use]
+    pub fn query(&self) -> QueryBuilder<Initial> {
+        QueryBuilder::new()
+    }
+
+    /// Execute a query and return the results.
+    ///
+    /// This method plans and executes the query using the hybrid query planner.
+    /// The planner applies optimization rules and chooses the best execution
+    /// strategy based on cost estimation.
+    ///
+    /// # Arguments
+    ///
+    /// * `query` - The query to execute
+    ///
+    /// # Example
+    ///
+    /// ```ignore
+    /// let query = db.query()
+    ///     .start(alice_id)
+    ///     .traverse("KNOWS")
+    ///     .rank_by_similarity(&embedding, 10)
+    ///     .build();
+    ///
+    /// let results = db.execute_query(query)?;
+    /// for row in results {
+    ///     println!("{:?}", row);
+    /// }
+    /// ```
+    pub fn execute_query(&self, query: Query) -> Result<QueryResults> {
+        #[cfg(feature = "observability")]
+        let _span = tracing::info_span!("execute_query").entered();
+
+        // Use cached statistics for cost-based optimization
+        // Statistics are shared across all queries for this database instance
+        let planner = QueryPlanner::new(Arc::clone(&self.stats), Arc::clone(&self.current));
+        let physical_plan = planner.plan(query)?;
+
+        // Execute the plan
+        let executor = QueryExecutor::new(Arc::clone(&self.current), Arc::clone(&self.historical));
+
+        executor.execute(physical_plan)
+    }
+
+    /// Traverse from a node and rank results by similarity to an embedding.
+    ///
+    /// This is a convenience method for a common hybrid query pattern:
+    /// "Find nodes connected to X that are similar to Y."
+    ///
+    /// # Arguments
+    ///
+    /// * `source` - The starting node for traversal
+    /// * `edge_label` - Edge type to traverse (e.g., "KNOWS")
+    /// * `embedding` - Target embedding to rank by similarity
+    /// * `k` - Maximum number of results to return
+    ///
+    /// # Example
+    ///
+    /// ```ignore
+    /// // "Who does Alice know that's similar to Bob?"
+    /// let results = db.traverse_and_rank(
+    ///     alice_id,
+    ///     "KNOWS",
+    ///     &bob_embedding,
+    ///     10
+    /// )?;
+    ///
+    /// for row in results {
+    ///     println!("Found: {:?}", row.node_id);
+    /// }
+    /// ```
+    pub fn traverse_and_rank(
+        &self,
+        source: NodeId,
+        edge_label: &str,
+        embedding: &[f32],
+        k: usize,
+    ) -> Result<QueryResults> {
+        #[cfg(feature = "observability")]
+        let _span = tracing::info_span!("traverse_and_rank").entered();
+
+        let query = self
+            .query()
+            .start(source)
+            .traverse(edge_label)
+            .rank_by_similarity(embedding, k)
+            .build();
+
+        self.execute_query(query)
+    }
+
+    /// Find similar nodes at a specific point in time.
+    ///
+    /// This is a convenience method for temporal vector queries:
+    /// "What was similar to this embedding at time T?"
+    ///
+    /// # Arguments
+    ///
+    /// * `embedding` - Query embedding
+    /// * `k` - Maximum number of results
+    /// * `valid_time` - Valid time for the query
+    /// * `transaction_time` - Transaction time for the query
+    ///
+    /// # Example
+    ///
+    /// ```ignore
+    /// // "What concepts were similar to this in 2023?"
+    /// let results = db.find_similar_at_time(
+    ///     &query_embedding,
+    ///     10,
+    ///     timestamp_2023,
+    ///     timestamp_2023
+    /// )?;
+    /// ```
+    pub fn find_similar_at_time(
+        &self,
+        embedding: &[f32],
+        k: usize,
+        valid_time: Timestamp,
+        transaction_time: Timestamp,
+    ) -> Result<QueryResults> {
+        #[cfg(feature = "observability")]
+        let _span = tracing::info_span!("find_similar_at_time").entered();
+
+        let query = self
+            .query()
+            .as_of(valid_time, transaction_time)
+            .find_similar(embedding, k)
+            .build();
+
+        self.execute_query(query)
+    }
+
+    /// Execute a full hybrid query combining graph, vector, and temporal.
+    ///
+    /// This is a convenience method for the most complex query pattern:
+    /// "Who did X know at time T that was similar to Y?"
+    ///
+    /// # Arguments
+    ///
+    /// * `source` - Starting node for traversal
+    /// * `edge_label` - Edge type to traverse
+    /// * `embedding` - Target embedding to rank by similarity
+    /// * `k` - Maximum number of results
+    /// * `valid_time` - Valid time for the query
+    /// * `transaction_time` - Transaction time for the query
+    ///
+    /// # Example
+    ///
+    /// ```ignore
+    /// // "Who did Alice know in 2023 that was similar to Bob?"
+    /// let results = db.traverse_and_rank_at_time(
+    ///     alice_id,
+    ///     "KNOWS",
+    ///     &bob_embedding,
+    ///     10,
+    ///     timestamp_2023,
+    ///     timestamp_2023
+    /// )?;
+    /// ```
+    pub fn traverse_and_rank_at_time(
+        &self,
+        source: NodeId,
+        edge_label: &str,
+        embedding: &[f32],
+        k: usize,
+        valid_time: Timestamp,
+        transaction_time: Timestamp,
+    ) -> Result<QueryResults> {
+        #[cfg(feature = "observability")]
+        let _span = tracing::info_span!("traverse_and_rank_at_time").entered();
+
+        let query = self
+            .query()
+            .as_of(valid_time, transaction_time)
+            .start(source)
+            .traverse(edge_label)
+            .rank_by_similarity(embedding, k)
+            .build();
+
+        self.execute_query(query)
+    }
+
+    /// Execute a Cypher query.
+    ///
+    /// This is a placeholder for future Cypher support.
+    #[cfg(feature = "cypher")]
+    pub fn cypher(&self, _query: &str) -> Result<QueryResults> {
+        use crate::query::executor::{QueryRow, ResultIterator};
+
+        // Minimal empty iterator to satisfy the return type
+        struct EmptyCypherIterator;
+        impl ResultIterator for EmptyCypherIterator {
+            fn next(&mut self) -> Option<Result<QueryRow>> {
+                None
+            }
+            fn size_hint(&self) -> (usize, Option<usize>) {
+                (0, Some(0))
+            }
+        }
+
+        Ok(QueryResults::new(Box::new(EmptyCypherIterator)))
+    }
+
+    /// Execute a Cypher query with parameters.
+    ///
+    /// This is a placeholder for future Cypher support.
+    #[cfg(feature = "cypher")]
+    pub fn cypher_with_params(
+        &self,
+        query: &str,
+        _params: crate::core::property::PropertyMap,
+    ) -> Result<QueryResults> {
+        self.cypher(query)
+    }
+}
